requirements.txt
Orange3
statsmodels
pandas
scipy
numpy
